updates  - copied profile.vue to patientProfile.vue, deleted profile.vue  - updated navbar to link patient to patient search, removed extra link  - updated patientSearch.vue to not stop halfway on the page  - switched to navigateTo
